% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allPre500.R
\docType{data}
\name{allPre500}
\alias{allPre500}
\title{The dataset of the prediction result for some days for 2014 Shandong, interpolated by constrained optimization.}
\format{DataFrame
\describe{
\item{dk}{predicted (by ensemble mixed model) or interpolated (by constrained optimization) estimate for rowname id and the kth year of day}
\item{row name}{location id, corresponding to the raster id}
#'
}}
\source{
Collected
}
\usage{
allPre500
}
\description{
The dataset of the prediction result for some days for 2014 Shandong, interpolated by constrained optimization.
}
\examples{
 allPre500
}
\keyword{datasets}
