% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnd.R
\docType{data}
\name{bnd}
\alias{bnd}
\title{BND spatial topology data for use in spatial effect modeling.}
\format{BND format for use in BayesX package
\describe{
#'List data object
}}
\source{
Collected
}
\usage{
bnd
}
\description{
BND spatial topology data for use in spatial effect modeling.
}
\examples{
  bnd
}
\keyword{datasets}
