% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countylayer.R
\docType{data}
\name{countylayer}
\alias{countylayer}
\title{County layer map for illustration of block kriging.}
\format{SpatialPolygonDataFrame
\describe{
\item{ID_0}{Privence name}
\item{NAME_3}{county name}
\item{...}{...}
#'
}}
\source{
Collected
}
\usage{
countylayer
}
\description{
County layer map for illustration of block kriging.
}
\examples{
  countylayer
}
\keyword{datasets}
