\name{extractVTIF}
\alias{extractVTIF}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Extract GeoTiff Data}
\description{Extract geotiff image.}
\usage{
extractVTIF(tarshp, tifRaster)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{tarshp}{spatial point object, data format: \code{PointDataFrame}}
  \item{tifRaster}{tifRaster object}
}
\value{ values for extracted. Format: \code{vector}}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
