% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtifRst.R
\docType{data}
\name{gtifRst}
\alias{gtifRst}
\title{The 2014 time series of PM2.5 concentrations of Shandong province, with many missing values.}
\format{Raster
\describe{
GeoTiff format image of AOD to demonstrate use of extractVTIF
}}
\source{
NASA MAIMIC data
}
\usage{
gtifRst
}
\description{
The 2014 time series of PM2.5 concentrations of Shandong province, with many missing values.
}
\examples{
  gtifRst
}
\keyword{datasets}
