\name{points2Raster}
\alias{points2Raster}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Generation of Grid Surface Using the predicted/Interpolated Values}
\description{This combines the predicted values's output with the corresponding spatial point data frame to generate the grid surface.  Please use this function with the output spatial point data frame generated by genRaster }
\usage{
points2Raster(spoints, tarVar, dx = 2000, dy = 2000)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{spoints}{Spatial point data frame. This data frame is based on the output by the function, genRaster with its predicted or interpolated value  field. }
  \item{tarVar}{Field name such as pollutant concentration used to make the grid.}
  \item{dx}{Size of resolution along x coordinate}
  \item{dy}{Size of resolution along y coordinate}
}
\value{Convert the points into Raster }

%% ~Make other sections like Warning with \section{Warning }{....} ~

\examples{
  data("spointspre")
  praster=points2Raster(spointspre,"pre_m",dx=2000,dy=2000)
  raster::plot(praster)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
