% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pol_season_trends.R
\docType{data}
\name{pol_season_trends}
\alias{pol_season_trends}
\title{pol_season_trends .}
\format{DataFrame
\describe{
\item{date}{date}
\item{pv1}{1st temporal basis function}
\item{pv2}{2st temporal basis function}
\item{tid}{unique temporal identifier}
#'
}}
\source{
simulated
}
\usage{
pol_season_trends
}
\description{
pol_season_trends. Temoral basis function output of log PM2.5 for Shan dong 2014 data
}
\examples{
  pol_season_trends
}
\keyword{datasets}
