% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prnside.R
\docType{data}
\name{prnside}
\alias{prnside}
\title{Side to limit the Thiessen's polygons.}
\format{SpatialPolygonsDataFrame
\describe{
\item{AREA}{AREA}
\item{PERIMETER}{PERIMETER}
\item{BOU2_4M_}{BOU2_4M_}
\item{BOU2_4M_ID}{BOU2_4M_ID}
\item{ADCODE93}{ADCODE93}
\item{ADCODE99}{ADCODE99}
}}
\source{
Collected
}
\usage{
prnside
}
\description{
Side shape file
}
\examples{
  prnside
}
\keyword{datasets}
