% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplepnt.R
\docType{data}
\name{samplepnt}
\alias{samplepnt}
\title{Sample data for generation of Thiessen polygons.}
\format{SpatialPointDataFrame
\describe{
\item{provence}{Privence name}
\item{city}{city name}
\item{geocode}{geocode}
\item{name}{name}
\item{code}{code}
\item{x}{x}
\item{y}{y}
\item{rid}{region id}
#'
}}
\source{
Collected
}
\usage{
samplepnt
}
\description{
Sample data for generation of Thiessen polygons.
}
\examples{
  samplepnt
}
\keyword{datasets}
