% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shd140401pcovs.R
\docType{data}
\name{shd140401pcovs}
\alias{shd140401pcovs}
\title{The dataset of 04/01/2014 prediction dataset for the raster spoint_pre covering the Shandong with 2km x 2km grid .}
\format{SpatialPointDataFrame
\describe{
\item{layer}{}
\item{gid}{unique location identifier}
\item{rid}{region id}
\item{ndvi}{ndvi}
\item{aod}{aod}
\item{wnd_avg}{wind speed}
\item{monthAv}{regional monthy average }
}}
\source{
Collected
}
\usage{
shd140401pcovs
}
\description{
The dataset of 04/01/2014 prediction dataset for the raster spoint_pre covering the Shandong with 2km x 2km grid .
}
\examples{
 shd140401pcovs

}
\keyword{datasets}
