% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shdSeries2014.R
\docType{data}
\name{shdSeries2014}
\alias{shdSeries2014}
\title{The 2014 time series of PM2.5 concentrations of Shandong province,  with missing approach.}
\format{DataFrame
\describe{
\item{siteid}{site id for monitoring station}
\item{date}{monitoring date}
\item{obs}{average over the hourly observed values or imputed value by SVD}
#'
}}
\source{
Collected
}
\usage{
shdSeries2014
}
\description{
The 2014 time series of PM2.5 concentrations of Shandong province,  with missing approach.
}
\examples{
  shdSeries2014
}
\keyword{datasets}
