% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainsample.R
\docType{data}
\name{trainsample}
\alias{trainsample}
\title{The dataset of 2014 training sample for the Shandong with missing values imputed using SVD.}
\format{\code{DataFrame}
\describe{
\item{provence}{provence}
\item{city}{city}
\item{geocode}{geocode}
\item{name}{name}
\item{code}{code}
\item{x}{x}
\item{y}{y}
\item{siteid }{siteid}
\item{rid }{rid}
\item{ndvi }{ndvi}
\item{aod }{aod}
\item{wnd_avg}{wind speed}
\item{monthAv}{regional monthy average }
\item{date}{date}
\item{month }{month}
\item{pm25 }{pm2.5}
\item{logpm25}{log PM2.5}
\item{tid}{tid}
}}
\source{
Collected
}
\usage{
trainsample
}
\description{
The dataset of 2014 training sample for the Shandong with missing values imputed using SVD.
}
\examples{
 trainsample

}
\keyword{datasets}
